requirements.txt
ruuvitag_sensor==0.12.0
paho_mqtt==1.4.0
PyYAML==3.13
